Implementation of shopping cart

+++ src/main/java/pl/koder95/bso/service/impl/UserServiceImpl.java
@@ -1,16 +1,15 @@
 package pl.koder95.bso.service.impl;
 
 import java.util.Optional;
-import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import pl.koder95.bso.dto.UserRegistrationRequestDto;
 import pl.koder95.bso.dto.UserResponseDto;
 import pl.koder95.bso.exception.EntityNotFoundException;
 import pl.koder95.bso.exception.RegistrationException;
 import pl.koder95.bso.mapper.UserMapper;
-import pl.koder95.bso.model.Role;
 import pl.koder95.bso.model.RoleName;
 import pl.koder95.bso.model.User;
 import pl.koder95.bso.repository.RoleRepository;
@@ -26,6 +25,7 @@ public class UserServiceImpl implements UserService {
     private final PasswordEncoder passwordEncoder;
 
     @Override
+    @Transactional
     public UserResponseDto register(UserRegistrationRequestDto requestDto)
             throws RegistrationException {
         if (userRepository.existsByEmail(requestDto.getEmail())) {
@@ -34,7 +34,7 @@ public class UserServiceImpl implements UserService {
             );
         }
         User model = userMapper.toModel(requestDto);
-        model.setRoles(Set.of(roleRepository.findByName(RoleName.ROLE_USER)));
+        model.getRoles().add(roleRepository.findByName(RoleName.ROLE_USER));
         model.setPassword(passwordEncoder.encode(model.getPassword()));
         User saved = userRepository.save(model);
         return userMapper.toDto(saved);
@@ -48,12 +48,19 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void delete(Long id) {
-        userRepository.deleteById(id);
+    @Transactional
+    public void deleteById(Long id) {
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isPresent()) {
+            userRepository.deleteById(id);
+        }
+        optionalUser.orElseThrow(() -> new EntityNotFoundException("User not found"));
     }
 
     @Override
-    public Optional<UserResponseDto> findByEmail(String email) {
-        return userRepository.findByEmail(email).map(userMapper::toDto);
+    public UserResponseDto findByEmail(String email) {
+        return userRepository.findByEmail(email).map(userMapper::toDto).orElseThrow(
+                () -> new EntityNotFoundException("User not found")
+        );
     }
 }
